tidy up chat UI layout
