+++ packages/frontend/editor-ui/src/features/chatHub/ChatView.vue
@@ -2,23 +2,15 @@
 import { ref, computed, watch, nextTick, onMounted, useTemplateRef } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
 import { v4 as uuidv4 } from 'uuid';
-import hljs from 'highlight.js/lib/core';
 
-import { N8nHeading, N8nIcon, N8nText, N8nScrollArea, N8nIconButton } from '@n8n/design-system';
-import PageViewLayout from '@/components/layouts/PageViewLayout.vue';
+import { N8nIcon, N8nScrollArea, N8nIconButton } from '@n8n/design-system';
 import ModelSelector from './components/ModelSelector.vue';
 import CredentialSelectorModal from './components/CredentialSelectorModal.vue';
 
 import { useChatStore } from './chat.store';
-import { useUsersStore } from '@/stores/users.store';
 import { useCredentialsStore } from '@/stores/credentials.store';
 import { useUIStore } from '@/stores/ui.store';
-import {
-	credentialsMapSchema,
-	type ChatMessage,
-	type CredentialsMap,
-	type Suggestion,
-} from './chat.types';
+import { credentialsMapSchema, type CredentialsMap, type Suggestion } from './chat.types';
 import {
 	chatHubConversationModelSchema,
 	type ChatHubProvider,
@@ -26,31 +18,33 @@ import {
 	type ChatHubConversationModel,
 	chatHubProviderSchema,
 } from '@n8n/api-types';
-import VueMarkdown from 'vue-markdown-render';
-import markdownLink from 'markdown-it-link-attributes';
-import type MarkdownIt from 'markdown-it';
-import { useLocalStorage } from '@vueuse/core';
+import { useLocalStorage, useMediaQuery } from '@vueuse/core';
 import {
 	LOCAL_STORAGE_CHAT_HUB_SELECTED_MODEL,
 	LOCAL_STORAGE_CHAT_HUB_CREDENTIALS,
 	CHAT_HUB_SIDE_MENU_DRAWER_MODAL_KEY,
 } from '@/constants';
-import { CHAT_CONVERSATION_VIEW, CHAT_VIEW, SUGGESTIONS } from '@/features/chatHub/constants';
+import {
+	CHAT_CONVERSATION_VIEW,
+	CHAT_VIEW,
+	MOBILE_MEDIA_QUERY,
+} from '@/features/chatHub/constants';
 import { findOneFromModelsResponse } from '@/features/chatHub/chat.utils';
 import { useToast } from '@/composables/useToast';
-import { useDeviceSupport } from '@n8n/composables/useDeviceSupport';
+import ChatMessage from '@/features/chatHub/components/ChatMessage.vue';
+import ChatPrompt from '@/features/chatHub/components/ChatPrompt.vue';
+import ChatTypingIndicator from '@/features/chatHub/components/ChatTypingIndicator.vue';
+import ChatStarter from '@/features/chatHub/components/ChatStarter.vue';
 
 const router = useRouter();
 const route = useRoute();
 const chatStore = useChatStore();
-const userStore = useUsersStore();
 const credentialsStore = useCredentialsStore();
 const uiStore = useUIStore();
 const toast = useToast();
-const { isMobileDevice } = useDeviceSupport();
+const isMobileDevice = useMediaQuery(MOBILE_MEDIA_QUERY);
 
 const inputRef = useTemplateRef('inputRef');
-const message = ref('');
 const sessionId = computed<string>(() =>
 	typeof route.params.id === 'string' ? route.params.id : uuidv4(),
 );
@@ -230,8 +224,8 @@ function onCreateNewCredential(provider: ChatHubProvider) {
 	uiStore.openNewCredential(PROVIDER_CREDENTIAL_TYPE_MAP[provider]);
 }
 
-function onSubmit() {
-	if (!message.value.trim() || chatStore.isResponding || !selectedModel.value) {
+function onSubmit(message: string) {
+	if (!message.trim() || chatStore.isResponding || !selectedModel.value) {
 		return;
 	}
 
@@ -241,14 +235,14 @@ function onSubmit() {
 		return;
 	}
 
-	chatStore.askAI(sessionId.value, message.value, selectedModel.value, {
+	chatStore.askAI(sessionId.value, message, selectedModel.value, {
 		[PROVIDER_CREDENTIAL_TYPE_MAP[selectedModel.value.provider]]: {
 			id: credentialsId,
 			name: '',
 		},
 	});
 
-	message.value = '';
+	inputRef.value?.setText('');
 
 	if (isNewSession.value) {
 		void router.push({ name: CHAT_CONVERSATION_VIEW, params: { id: sessionId.value } });
@@ -256,42 +250,21 @@ function onSubmit() {
 }
 
 function onSuggestionClick(s: Suggestion) {
-	message.value = `${s.title} ${s.subtitle}`;
+	inputRef.value?.setText(`${s.title} ${s.subtitle}`);
 }
-
-function onAttach() {}
-
-function onMic() {}
-
-function messageText(msg: ChatMessage) {
-	return msg.type === 'message' ? msg.text : `**Error:** ${msg.content}`;
-}
-
-const markdownOptions = {
-	highlight(str: string, lang: string) {
-		if (lang && hljs.getLanguage(lang)) {
-			try {
-				return hljs.highlight(str, { language: lang }).value;
-			} catch {}
-		}
-
-		return ''; // use external default escaping
-	},
-};
-
-const linksNewTabPlugin = (vueMarkdownItInstance: MarkdownIt) => {
-	vueMarkdownItInstance.use(markdownLink, {
-		attrs: {
-			target: '_blank',
-			rel: 'noopener',
-		},
-	});
-};
 </script>
 
 <template>
-	<PageViewLayout>
-		<div :class="$style.modelSelector">
+	<N8nScrollArea
+		ref="scrollAreaRef"
+		type="hover"
+		:enable-vertical-scroll="true"
+		:enable-horizontal-scroll="false"
+		:viewport-class="$style.scrollViewport"
+		as-child
+		:class="{ [$style.hasMessages]: hasMessages, [$style.isMobileDevice]: isMobileDevice }"
+	>
+		<div :class="$style.floating">
 			<N8nIconButton
 				v-if="isMobileDevice"
 				type="secondary"
@@ -307,6 +280,7 @@ const linksNewTabPlugin = (vueMarkdownItInstance: MarkdownIt) => {
 				@configure="onConfigure"
 			/>
 		</div>
+
 		<CredentialSelectorModal
 			v-if="credentialSelectorProvider"
 			:key="credentialSelectorProvider"
@@ -315,180 +289,51 @@ const linksNewTabPlugin = (vueMarkdownItInstance: MarkdownIt) => {
 			@select="onCredentialSelected"
 			@create-new="onCreateNewCredential"
 		/>
-		<div
-			:class="{
-				[$style.content]: true,
-				[$style.centered]: !hasMessages,
-			}"
-		>
-			<section
-				:class="{
-					[$style.section]: true,
-					[$style.fullHeight]: hasMessages,
-				}"
-			>
-				<div v-if="!hasMessages" :class="$style.header">
-					<N8nHeading tag="h2" bold size="xlarge">
-						{{
-							`Good morning, ${userStore.currentUser?.firstName ?? userStore.currentUser?.fullName ?? 'User'}!`
-						}}
-					</N8nHeading>
-				</div>
 
-				<div v-if="!hasMessages" :class="$style.suggestions">
-					<button
-						v-for="s in SUGGESTIONS"
-						:key="s.title"
-						type="button"
-						:class="$style.card"
-						@click="onSuggestionClick(s)"
-					>
-						<div :class="$style.cardIcon" aria-hidden="true">
-							<N8nText size="xlarge">{{ s.icon }}</N8nText>
-						</div>
-						<div :class="$style.cardText">
-							<N8nText bold color="text-dark">{{ s.title }}</N8nText>
-							<N8nText color="text-base">{{ s.subtitle }}</N8nText>
-						</div>
-					</button>
-				</div>
+		<div :class="$style.scrollable">
+			<ChatStarter
+				v-if="!hasMessages"
+				:class="$style.starter"
+				:is-mobile-device="isMobileDevice"
+				@select="onSuggestionClick"
+			/>
 
-				<!-- Chat thread -->
-				<template v-else>
-					<div :class="$style.threadContainer">
-						<N8nScrollArea
-							ref="scrollAreaRef"
-							type="hover"
-							:enable-vertical-scroll="true"
-							:enable-horizontal-scroll="false"
-							:class="$style.threadWrap"
-						>
-							<div ref="messagesRef" :class="$style.thread" role="log" aria-live="polite">
-								<div
-									v-for="m in chatMessages"
-									:key="m.id"
-									:class="[
-										$style.message,
-										m.role === 'user' ? $style.user : $style.assistant,
-										m.type === 'error' && $style.error,
-									]"
-								>
-									<div :class="$style.avatar">
-										<N8nIcon
-											:icon="m.role === 'user' ? 'user' : 'sparkles'"
-											width="20"
-											height="20"
-										/>
-									</div>
-									<div
-										:class="{
-											[$style.chatMessage]: true,
-											[$style.chatMessageFromUser]: m.role === 'user',
-											[$style.chatMessageFromAssistant]: m.role === 'assistant',
-										}"
-									>
-										<VueMarkdown
-											:class="$style.chatMessageMarkdown"
-											:source="messageText(m)"
-											:options="markdownOptions"
-											:plugins="[linksNewTabPlugin]"
-										/>
-									</div>
-								</div>
+			<div v-else ref="messagesRef" role="log" aria-live="polite" :class="$style.messageList">
+				<ChatMessage v-for="m in chatMessages" :key="m.id" :message="m" :compact="isMobileDevice" />
 
-								<!-- Typing indicator while streaming -->
-								<div v-if="chatStore.isResponding" :class="[$style.message, $style.assistant]">
-									<div :class="$style.avatar">
-										<N8nIcon icon="sparkles" width="20" height="20" />
-									</div>
-									<div :class="$style.bubble">
-										<span :class="$style.typing"><i></i><i></i><i></i></span>
-									</div>
-								</div>
-							</div>
-						</N8nScrollArea>
+				<div v-if="chatStore.isResponding" :class="[$style.message, $style.assistant]">
+					<div :class="$style.avatar">
+						<N8nIcon icon="sparkles" width="20" height="20" />
 					</div>
-				</template>
-
-				<!-- Prompt -->
-				<form :class="$style.prompt" @submit.prevent="onSubmit">
-					<div :class="$style.inputWrap">
-						<input
-							ref="inputRef"
-							v-model="message"
-							:class="$style.input"
-							type="text"
-							:placeholder="inputPlaceholder"
-							autocomplete="off"
-							autofocus
-							:disabled="chatStore.isResponding"
-						/>
-
-						<div :class="$style.actions">
-							<button
-								:class="$style.iconBtn"
-								type="button"
-								title="Attach"
-								:disabled="chatStore.isResponding"
-								@click="onAttach"
-							>
-								<N8nIcon icon="paperclip" width="20" height="20" />
-							</button>
-							<button
-								:class="$style.iconBtn"
-								type="button"
-								title="Voice"
-								:disabled="chatStore.isResponding"
-								@click="onMic"
-							>
-								<N8nIcon icon="mic" width="20" height="20" />
-							</button>
-							<button
-								:class="$style.sendBtn"
-								type="submit"
-								:disabled="chatStore.isResponding || !message.trim()"
-							>
-								<span v-if="!chatStore.isResponding">Send</span>
-								<span v-else>…</span>
-							</button>
-						</div>
+					<div :class="$style.bubble">
+						<ChatTypingIndicator v-if="chatStore.isResponding" />
 					</div>
-					<N8nText :class="$style.disclaimer" color="text-light" size="small">
-						AI may make mistakes. Check important info.
-						<br />
-						{{ sessionId }}
-					</N8nText>
-				</form>
-			</section>
+				</div>
+			</div>
+
+			<div :class="$style.promptContainer">
+				<ChatPrompt
+					ref="inputRef"
+					:class="$style.prompt"
+					:placeholder="inputPlaceholder"
+					:disabled="chatStore.isResponding"
+					:session-id="sessionId"
+					@submit="onSubmit"
+				/>
+			</div>
 		</div>
-	</PageViewLayout>
+	</N8nScrollArea>
 </template>
 
 <style lang="scss" module>
-.content {
+.scrollable {
+	width: 100%;
+	min-height: 100%;
 	display: flex;
 	flex-direction: column;
-	gap: var(--spacing--md);
-	padding-bottom: var(--spacing--lg);
-
-	height: 100%;
-	min-height: 0;
-}
-
-.centered {
+	align-items: stretch;
 	justify-content: center;
-}
-
-.section {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: var(--spacing--lg);
-}
-
-.fullHeight {
-	flex: 1;
-	min-height: 0;
+	gap: var(--spacing--2xl);
 }
 
 .header {
@@ -497,243 +342,59 @@ const linksNewTabPlugin = (vueMarkdownItInstance: MarkdownIt) => {
 	align-items: center;
 }
 
-/* Suggestions */
-.suggestions {
-	display: grid;
-	grid-template-columns: repeat(2, minmax(220px, 1fr));
+.starter {
+	.isMobileDevice & {
+		padding-top: 100px;
+		padding-bottom: 200px;
+	}
+}
+
+.messageList {
+	max-width: 55rem;
+	min-height: 100vh;
+	align-self: center;
+	display: flex;
+	flex-direction: column;
 	gap: var(--spacing--md);
-	width: min(960px, 90%);
-	margin-top: var(--spacing--md);
-}
-@media (max-width: 800px) {
-	.suggestions {
-		grid-template-columns: 1fr;
+	padding-top: 100px;
+	padding-bottom: 200px;
+	padding-inline: 64px;
+
+	.isMobileDevice & {
+		padding-inline: var(--spacing--md);
 	}
 }
-.card {
+
+.promptContainer {
 	display: flex;
-	align-items: flex-start;
-	gap: var(--spacing--sm);
-	padding: var(--spacing--md);
-	border: 1px solid var(--color--foreground);
-	background: var(--color--background);
-	border-radius: var(--radius--lg);
-	text-align: left;
-	cursor: pointer;
-	transition:
-		transform 0.06s ease,
-		background 0.06s ease,
-		border-color 0.06s ease;
-}
-.card:hover {
-	border-color: var(--color--primary);
-	background: rgba(124, 58, 237, 0.04);
-}
-.cardIcon {
-	height: 100%;
-	display: flex;
-	align-items: center;
-}
-.cardText {
-	display: grid;
-	gap: 2px;
-}
+	justify-content: center;
 
-.threadWrap {
-	flex: 1;
-	height: 100%;
-	min-height: 0;
-}
-
-.threadContainer {
-	width: min(960px, 90%);
-	flex: 1;
-	min-height: 0;
-	display: flex;
-}
-
-.thread {
-	padding: var(--spacing--md);
-	background: var(--color--background--light-2);
-}
-
-.message {
-	display: grid;
-	grid-template-columns: 28px 1fr;
-	gap: var(--spacing--sm);
-	margin-bottom: var(--spacing--md);
-}
-.avatar {
-	display: grid;
-	place-items: center;
-	width: 28px;
-	height: 28px;
-	border-radius: 50%;
-	background: var(--color--background--light-3);
-	color: var(--color--text--tint-1);
-}
-
-.chatMessage {
-	display: block;
-	position: relative;
-	max-width: fit-content;
-	padding: var(--spacing--md);
-	border-radius: var(--radius--lg);
-
-	&.chatMessageFromAssistant {
-		background-color: var(--color--background);
-	}
-
-	&.chatMessageFromUser {
-		background-color: var(--color--background--shade-1);
-	}
-
-	> .chatMessageMarkdown {
-		display: block;
-		box-sizing: border-box;
-		font-size: inherit;
-
-		> *:first-child {
-			margin-top: 0;
-		}
-
-		> *:last-child {
-			margin-bottom: 0;
-		}
-
-		p {
-			margin: var(--spacing--xs) 0;
-		}
-
-		pre {
-			font-family: inherit;
-			font-size: inherit;
-			margin: 0;
-			white-space: pre-wrap;
-			box-sizing: border-box;
-			padding: var(--chat--spacing);
-			background: var(--chat--message--pre--background);
-			border-radius: var(--chat--border-radius);
-		}
+	.isMobileDevice &,
+	.hasMessages & {
+		position: absolute;
+		bottom: 0;
+		left: 0;
+		width: 100%;
+		padding-block: var(--spacing--md);
+		background: linear-gradient(transparent 0%, var(--color--background--light-2) 30%);
 	}
 }
 
-/* Typing indicator */
-.typing {
-	display: inline-flex;
-	gap: 6px;
-}
-.typing i {
-	width: 6px;
-	height: 6px;
-	border-radius: 50%;
-	background: currentColor;
-	opacity: 0.35;
-	animation: blink 1.2s infinite;
-}
-.typing i:nth-child(2) {
-	animation-delay: 0.2s;
-}
-.typing i:nth-child(3) {
-	animation-delay: 0.4s;
-}
-
-@keyframes blink {
-	0%,
-	80%,
-	100% {
-		opacity: 0.35;
-		transform: translateY(0);
-	}
-	40% {
-		opacity: 1;
-		transform: translateY(-2px);
-	}
-}
-
-/* Prompt */
 .prompt {
-	display: grid;
-	place-items: center;
 	width: 100%;
-	margin-top: var(--spacing--md);
-}
-.inputWrap {
-	position: relative;
-	display: flex;
-	align-items: center;
-	width: min(720px, 50vw);
-	min-width: 320px;
+	max-width: 55rem;
+	padding-inline: 64px;
 
-	& input:disabled {
-		cursor: not-allowed;
+	.isMobileDevice & {
+		padding-inline: var(--spacing--md);
 	}
 }
-.input {
-	flex: 1;
-	font: inherit;
-	padding: 14px 112px 14px 14px;
-	border: 1px solid var(--color--foreground);
-	background: var(--color--background--light-2);
-	color: var(--color--text--shade-1);
-	border-radius: 16px;
-	outline: none;
-}
-.input:focus {
-	border-color: var(--color--primary);
-	box-shadow: 0 0 0 3px rgba(124, 58, 237, 0.15);
-}
 
-/* Right-side actions */
-.actions {
+.floating {
 	position: absolute;
-	right: 6px;
-	top: 50%;
-	transform: translateY(-50%);
-	display: flex;
-	align-items: center;
-	gap: 6px;
-}
-.iconBtn {
-	display: grid;
-	place-items: center;
-	width: 32px;
-	height: 32px;
-	border-radius: 10px;
-	border: none;
-	background: transparent;
-	color: var(--color--text--tint-1);
-	cursor: pointer;
-}
-.iconBtn:hover {
-	background: rgba(0, 0, 0, 0.04);
-	color: var(--color--text--shade-1);
-}
-.sendBtn {
-	height: 32px;
-	padding: 0 10px;
-	border-radius: 10px;
-	border: none;
-	background: var(--color--primary);
-	color: #fff;
-	font-weight: 600;
-	cursor: pointer;
-}
-.sendBtn:disabled {
-	opacity: 0.5;
-	cursor: not-allowed;
-}
-
-.disclaimer {
-	margin-top: var(--spacing--xs);
-	color: var(--color--text--tint-2);
-	text-align: center;
-}
-
-.modelSelector {
-	position: absolute;
-	top: var(--spacing-xs);
-	left: var(--spacing-xs);
+	padding: var(--spacing--xs);
+	top: 0;
+	left: 0;
 	z-index: 100;
 	display: flex;
 	align-items: center;
