+++ packages/frontend/editor-ui/src/features/chatHub/components/ChatSidebarContent.vue
@@ -1,12 +1,15 @@
 <script setup lang="ts">
+import MainSidebarUserArea from '@/components/MainSidebarUserArea.vue';
 import { CHAT_HUB_SIDE_MENU_DRAWER_MODAL_KEY, VIEWS } from '@/constants';
 import { useChatStore } from '@/features/chatHub/chat.store';
 import { groupConversationsByDate } from '@/features/chatHub/chat.utils';
 import { CHAT_CONVERSATION_VIEW, CHAT_VIEW } from '@/features/chatHub/constants';
 import { useUIStore } from '@/stores/ui.store';
-import { N8nIcon, N8nIconButton, N8nMenuItem, N8nText } from '@n8n/design-system';
+import { N8nIcon, N8nIconButton, N8nMenuItem, N8nScrollArea, N8nText } from '@n8n/design-system';
 import { computed, onMounted } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
+
+defineProps<{ isMobileDevice: boolean }>();
 
 const route = useRoute();
 const router = useRouter();
@@ -40,20 +43,25 @@ onMounted(async () => {
 </script>
 
 <template>
-	<div>
+	<div :class="[$style.component, { [$style.isMobileDevice]: isMobileDevice }]">
 		<div :class="$style.header">
-			<div :class="$style.returnButton" @click="onReturn">
-				<i>
-					<N8nIcon icon="arrow-left" />
-				</i>
+			<div :class="$style.returnButton" role="button" @click="onReturn">
+				<N8nIcon icon="arrow-left" />
 				<N8nText bold>Chat</N8nText>
 			</div>
-			<N8nIconButton title="New chat" icon="square-pen" type="tertiary" text @click="onNewChat" />
+			<N8nIconButton
+				title="New chat"
+				icon="square-pen"
+				type="tertiary"
+				text
+				:size="isMobileDevice ? 'large' : 'medium'"
+				@click="onNewChat"
+			/>
 		</div>
-		<div :class="['side-menu', $style.container]">
+		<N8nScrollArea as-child>
 			<div :class="$style.items">
 				<div v-for="group in groupedConversations" :key="group.group" :class="$style.group">
-					<N8nText :class="$style.groupHeader" size="xsmall" bold color="text-light">
+					<N8nText :class="$style.groupHeader" size="small" bold color="text-light">
 						{{ group.group }}
 					</N8nText>
 					<N8nMenuItem
@@ -69,25 +77,35 @@ onMounted(async () => {
 					/>
 				</div>
 			</div>
-		</div>
+		</N8nScrollArea>
+		<MainSidebarUserArea :fully-expanded="true" :is-collapsed="false" />
 	</div>
 </template>
 
 <style lang="scss" module>
+.component {
+	display: flex;
+	flex-direction: column;
+	align-items: stretch;
+}
+
 .header {
 	display: flex;
 	align-items: center;
 	justify-content: space-between;
-	padding: var(--spacing-xs);
-	gap: var(--spacing-2xs);
+	padding: var(--spacing--xs);
+	gap: var(--spacing--2xs);
 }
 
 .returnButton {
 	cursor: pointer;
+	flex-grow: 1;
+	flex-shrink: 1;
 	display: flex;
-	gap: var(--spacing-3xs);
+	gap: var(--spacing--3xs);
 	align-items: center;
 	flex: 1;
+
 	&:hover {
 		color: var(--color--primary);
 	}
@@ -96,8 +114,12 @@ onMounted(async () => {
 .items {
 	display: flex;
 	flex-direction: column;
-	padding: 0 var(--spacing-3xs);
-	gap: var(--spacing-xs);
+	padding: 0 var(--spacing--3xs);
+	gap: var(--spacing--xs);
+
+	.isMobileDevice & {
+		gap: var(--spacing--sm);
+	}
 }
 
 .group {
@@ -106,12 +128,12 @@ onMounted(async () => {
 }
 
 .groupHeader {
-	padding: 0 var(--spacing-3xs) var(--spacing-3xs) var(--spacing-3xs);
+	padding: 0 var(--spacing--3xs) var(--spacing--3xs) var(--spacing--3xs);
 }
 
 .loading,
 .empty {
-	padding: var(--spacing-xs);
+	padding: var(--spacing--xs);
 	text-align: center;
 }
 </style>
