+++ packages/frontend/editor-ui/src/features/chatHub/constants.ts
@@ -35,3 +35,5 @@ export const providerDisplayNames: Record<ChatHubProvider, string> = {
 	anthropic: 'Anthropic',
 	google: 'Google',
 };
+
+export const MOBILE_MEDIA_QUERY = '(max-width: 768px)';
